Added confirmation boxes to deleting user accounts and teams

+++ src/components/AccountManagement/DeleteAccountButton.jsx
@@ -10,15 +10,19 @@ export default function DeleteAccountButton() {
 
     const handleDeleteAccount = async () => {
         console.log("Attempting to Delete Account")
-        const res = await axios.delete('/api/users/delete', { data: { userId: userId }});
-        if (res.data.success) {
-            console.log("Succeeded Delete Account");
-            dispatch({ type: 'SET_USERINFO', payload: {email: null, userId: null}});
-            navigate('/');
-        }
-        else {
-            console.log(res.data);
-            console.log("Failed Delete");
+
+        const verified = confirm("Are you sure you want to delete this account? It can't be recovered if you do!");
+        if (verified) {
+            const res = await axios.delete('/api/users/delete', { data: { userId: userId }});
+            if (res.data.success) {
+                console.log("Succeeded Delete Account");
+                dispatch({ type: 'SET_USERINFO', payload: {email: null, userId: null}});
+                navigate('/');
+            }
+            else {
+                console.log(res.data);
+                console.log("Failed Delete");
+            }
         }
     }
